gui-api/requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.4
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
iniconfig==1.1.1
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.7
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.2
python-dotenv==0.15.0
pyzmq==20.0.0
six==1.15.0
toml==0.10.2
virtualenv==20.2.2
Werkzeug==1.0.1
wrapt==1.12.1
